requirements.txt
attrs==21.4.0
Cerberus==1.3.4
click==8.0.3
coverage==6.3.2
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
PyYAML==6.0
tomli==2.0.1
